Create imageio_utility.h
with parts of ktxtools/utility.h useful in imageio.

Move image.hpp to imageio.
