+++ tools/ktx/command_extract.cpp
@@ -527,6 +527,8 @@ void CommandExtract::decodeAndSaveASTC(std::string filepath, bool appendExtensio
             height);
 }
 
+using namespace imageio;
+
 void CommandExtract::unpackAndSave422(std::string filepath, bool appendExtension,
         VkFormat vkFormat, const FormatDescriptor& format, uint32_t width, uint32_t height,
         const char* data, std::size_t size) {
